requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.35.0
httpx==0.25.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
python-multipart==0.0.6
google-ads>=25.1.0
pandas==2.2.0
openpyxl==3.1.2
